Stepped back to 1.4 compliance to get rid of warnings.
git-svn-id: file:///home/igor/dev/stuff/svnbackup/trunk@532 ef7698a4-5110-0410-9fc6-c7eb3693863f

+++ wicket-contrib-database/src/java/wicket/contrib/database/DatabaseWebPage.java
@@ -103,7 +103,7 @@ public class DatabaseWebPage extends WebPage
 	
 	protected void delete(Class c, Long id)
 	{
-		if (id != -1)
+		if (id.longValue() != -1)
 		{
 			getDatabaseSession().delete(c, id);
 		}
